+++ src/Loader.php
@@ -5,33 +5,30 @@ namespace Rougin\Credo;
 /**
  * Loader
  *
- * An extension of the Loader Class of CodeIgniter.
- *
  * @package Credo
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
 class Loader extends \CI_Loader
 {
     /**
-     * Repository Loader
-     *
      * Loads and instantiates Doctrine-based repositories.
-     * It's designed to be called from application controllers.
      *
-     * @param   string|array $repository
-     * @return  object
+     * @param  string|array $repository
+     * @return self
      */
     public function repository($repository)
     {
-        if (is_array($repository)) {
-            foreach ($repository as $key => $value) {
-                $this->repository($value);
-            }
+        $repositories = (array) $repository;
 
-            return $this;
+        $path = APPPATH . 'repositories/';
+
+        foreach ($repositories as $repository) {
+            $file = ucfirst($repository);
+
+            $suffix = '_repository.php';
+
+            require $path . $file . $suffix;
         }
-
-        require APPPATH . 'repositories/' . ucfirst($repository) . '_repository.php';
 
         return $this;
     }